Move k8s resource pool methods to correct file
